# internal/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Dockmesh API
  summary: Container management platform — HTTP API
  description: |
    REST API for Dockmesh. All endpoints are versioned under `/api/v1`.

    ## Authentication

    Two bearer-token flavours are accepted on the `Authorization` header:

    - **User JWTs** — obtained via `POST /auth/login`, short-lived (15 min),
      refreshable via `POST /auth/refresh`. Use for interactive UI / CLI sessions.
    - **API tokens** — prefix `dmt_`, long-lived, created under
      Settings → API Tokens. Use for CI, scripts, external integrations.

    Both share the same permission model — a token's role dictates what it can do.

    ## Versioning

    Version `v1` is stable in the sense that we don't remove fields or change
    response shapes without warning. New fields may be added at any time, so
    clients should ignore unknown fields gracefully.

    ## Errors

    Non-2xx responses use a uniform shape: `{"error": "<human-readable message>"}`.
    HTTP status carries the category (400/401/403/404/409/500).

    ## Maintenance

    This spec is hand-maintained. `TestOpenAPIDriftAgainstRoutes` compares the
    registered chi routes against the operations declared here and fails CI on
    mismatch — so every PR that touches a route must also touch this file.
  version: "1.0.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Dockmesh
    url: https://dockmesh.dev
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /agents/{id}/upgrade:
    post:
      operationId: agents.upgrade
      tags: [agents]
      summary: Trigger a self-upgrade on a connected agent
      description: |
        The agent downloads a fresh `dockmesh-agent-linux-<arch>` binary
        from the server's `/install/` path and restarts its systemd unit.
        Agent must be currently connected.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Upgrade dispatched.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: upgrading }
                  version: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503":
          description: Agent is not currently connected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /stacks/{name}/discard:
    post:
      operationId: stacks.discard
      tags: [stacks]
      summary: Drop a ghost stack record without touching containers
      description: |
        Removes the deployment row + dependency edges for a stack whose
        compose file is gone from disk. Running containers are not
        touched — use this when the operator decides the stack should
        be forgotten by dockmesh but the workload should keep running
        (or will be cleaned up via `docker` directly).
      parameters:
        - $ref: "#/components/parameters/StackNamePath"
      responses:
        "204": { description: Discarded. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /roles/permissions:
    get:
      operationId: roles.permissions
      tags: [roles]
      summary: Catalogue of permissions that can be granted to a role
      description: |
        Returns all `Perm` values the server knows about with a
        human-readable description. Used by the role-create dialog to
        render the permission-picker.
      responses:
        "200":
          description: Permission catalogue.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [name, display_name, description, category, verb, danger_level]
                  properties:
                    name:
                      type: string
                      description: Backend identifier in `category.verb` dot-notation.
                      example: containers.exec
                    display_name:
                      type: string
                      description: Short human-readable label shown on the matrix cell tooltip and in Edit Role rows.
                      example: Shell into container
                    description:
                      type: string
                      description: One-line plain-language explanation. Renders as the `?` info-button tooltip.
                    category:
                      type: string
                      description: Resource group the permissions matrix renders this row under.
                      example: Containers
                    verb:
                      type: string
                      description: Matrix column header within the category — e.g., view / deploy / update / delete or per-resource sensitive verb.
                      example: exec
                    danger_level:
                      type: string
                      enum: [low, medium, high]
                      description: Drives UI tinting on dangerous grants.
                    sensitive:
                      type: boolean
                      description: When true the permission belongs in the matrix's "Sensitive operations" sub-section rather than the standard verb columns.
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT or dmt_<40-char-base64url>"
      description: |
        Accepts either a user session JWT (from `POST /auth/login`) or an API
        token (prefix `dmt_`, from Settings → API Tokens). The route layer
        checks role permissions after authentication.

  # Schemas used across multiple groups live here. Group-specific schemas
  # live inline under the operation to keep the spec readable when new
  # groups get added.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid credentials
...
  parameters:
...
    StackNamePath:
      name: name
      in: path
      description: Stack name — matches `[a-z0-9][a-z0-9-]*[a-z0-9]`, 2..63 chars.
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but the token's role lacks the required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...